requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Document parsing
pypdf==3.17.1
python-docx==0.8.11
python-pptx==0.6.21
markdown==3.5.1

# Embeddings & Vector Store
sentence-transformers==3.0.1
faiss-cpu==1.13.2

# Database
sqlalchemy==2.0.23

# LLM / Generation
openai==2.21.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
